Cleaned code and optimised for memory
Cleaned the Arduino code and added comments for better pick-up and read understanding.

Optimised for better memory efficency by using the correct data types. For example setReadingPrecisionCharacteristic was set to be of type Int when really it would be better to be of type Unsigned Char as the max value would never realistically be more than 10.